yolo5/requirements.txt
flask
pyyaml
loguru

# testing

pylint
pytest
pymongo

# aws boto3
boto3